@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
 	  <EnablePreviewFeatures>True</EnablePreviewFeatures>
 	  <IsPackable>false</IsPackable>
     <Authors>Olivier MARTY</Authors>
