@@ -28,10 +28,10 @@ namespace Utils.IO.Serialization
 
 		public NewWriter(Stream stream) : this(stream, new RawWriter().WriterDelegates) { }
 
-		public NewWriter(Stream stream, params IEnumerable<Delegate> converters)
-		{
-			this.Stream = stream ?? throw new ArgumentNullException(nameof(stream));
-			foreach (var converter in converters.Union(new RawWriter().WriterDelegates))
+                public NewWriter(Stream stream, params IEnumerable<Delegate> converters)
+                {
+                        this.Stream = stream ?? throw new ArgumentNullException(nameof(stream));
+                        foreach (var converter in converters.Union(new RawWriter().WriterDelegates))
 			{
 				var method = converter.GetMethodInfo();
 				var arguments = method.GetParameters();
@@ -44,7 +44,8 @@ namespace Utils.IO.Serialization
 			}
 		}
 
-		public NewWriter(Stream stream, params IEnumerable<IEnumerable<Delegate>> converters) : this(stream, converters.SelectMany(c => c)) { }
+                public NewWriter(Stream stream, params IEnumerable<IEnumerable<Delegate>> converters)
+                        : this(stream, converters.SelectMany(c => c)) { }
 
 
 		public void WriteByte(byte value) => Stream.WriteByte(value);
