@@ -26,10 +26,10 @@ namespace Utils.Files
 			string remainder = fullPath[root.Length..].TrimStart(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
 
 			// 4) Split on directory separators to get each segment
-			var segments = remainder.Split(
-				[Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar],
-				StringSplitOptions.RemoveEmptyEntries
-			);
+                        var segments = remainder.Split(
+                                [Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar],
+                                StringSplitOptions.RemoveEmptyEntries
+                        );
 
 			return (root, segments);
 		}
