@@ -365,8 +365,9 @@ public readonly struct Bytes :
 		if (string.IsNullOrWhiteSpace(s))
 			return Empty;
 
-		var values = s.Split([' ', '\n', '\t', '\r', ',', ';'],
-							 StringSplitOptions.RemoveEmptyEntries);
+                var values = s.Split(
+                        [' ', '\n', '\t', '\r', ',', ';'],
+                        StringSplitOptions.RemoveEmptyEntries);
 
 		var bytes = values.Select(v => byte.Parse(v, System.Globalization.NumberStyles.HexNumber))
 						  .ToArray();
